subir imágenes

+++ application/models/m_imagenes.php
@@ -100,7 +100,132 @@ class M_imagenes extends CI_Model {
 		$this->db->delete('ter_imagenes_on_albums');
 
 	}
+
+/*================================================================
+	Imágenes
+=================================================================*/
+
+	public function insertImagen($imagenNombre, $imagenFecha, $imagenLugar, $imagenDescripcion, $imagenRuta, $imagenAlbum) {
+
+		if ( strlen($imagenNombre) > 249 OR strlen($imagenNombre) == 0 ) {
+			die("Error en el envío del formulario");
+		} else {
+			$imagenNombre = addslashes( strip_tags($imagenNombre) );
+		}
+
+		if ( strlen($imagenFecha) > 12 ) {
+			die("Error en el envío del formulario");
+		} else {
+			$imagenFecha = addslashes( strip_tags($imagenFecha) );
+		}
+
+		if ( strlen($imagenLugar) > 249 OR strlen($imagenLugar) == 0 ) {
+			die("Error en el envío del formulario");
+		} else {
+			$imagenLugar = addslashes( strip_tags($imagenLugar) );
+		}		
+		
+		$imagenDescripcion = addslashes( $imagenDescripcion );
+
+		if ( strlen($imagenRuta) > 249 OR strlen($imagenRuta) == 0 ) {
+			die("Error en el envío del formulario");
+		} else {
+			$imagenRuta = addslashes( strip_tags($imagenRuta) );
+		}
+
+		if ( !is_numeric($imagenAlbum) ) {
+			die("Error en el envío del formulario");
+		}
+
+		$data = array(
+					   'nombre' 	 => $imagenNombre,
+					   'fecha' 		 => $imagenFecha,
+					   'lugar' 		 => $imagenLugar,
+					   'descripcion' => $imagenDescripcion,
+					   'ruta' 		 => $imagenRuta,
+					   'id_album'	 => $imagenAlbum
+					);
+
+		$query = $this->db->insert('ter_imagenes', $data);
+
+		$idImagen = $this->db->insert_id(); 
+		return $idImagen;
+		
+	} #insertImagen
 			
+	public function updateImagen($idImagen, $imagenNombre, $imagenFecha, $imagenLugar, $imagenDescripcion, $imagenRuta, $imagenAlbum) {
+
+		if ( !is_numeric($idImagen) ) {
+			die("Error en el envío del formulario");
+		}
+
+		if ( strlen($imagenNombre) > 249 OR strlen($imagenNombre) == 0 ) {
+			die("Error en el envío del formulario");
+		} else {
+			$imagenNombre = addslashes( strip_tags($imagenNombre) );
+		}
+
+		if ( strlen($imagenFecha) > 12 ) {
+			die("Error en el envío del formulario");
+		} else {
+			$imagenFecha = addslashes( strip_tags($imagenFecha) );
+		}
+
+		if ( strlen($imagenLugar) > 249 OR strlen($imagenLugar) == 0 ) {
+			die("Error en el envío del formulario");
+		} else {
+			$imagenLugar = addslashes( strip_tags($imagenLugar) );
+		}		
+		
+		$imagenDescripcion = addslashes( $imagenDescripcion );
+
+		if ( strlen($imagenRuta) > 249 OR strlen($imagenRuta) == 0 ) {
+			die("Error en el envío del formulario");
+		} else {
+			$imagenRuta = addslashes( strip_tags($imagenRuta) );
+		}
+
+		if ( !is_numeric($imagenAlbum) ) {
+			die("Error en el envío del formulario");
+		}
+
+		$data = array(
+			   'nombre' 	 => $imagenNombre,
+			   'fecha' 		 => $imagenFecha,
+			   'lugar' 		 => $imagenLugar,
+			   'descripcion' => $imagenDescripcion,
+			   'ruta' 		 => $imagenRuta,
+				'id_album'	 => $imagenAlbum
+			);
+
+		$this->db->where('id', $idImagen);
+		$this->db->update('ter_imagenes', $data);
+
+	}
+
+	public function getImagen($idImagen) {
+
+		$this->db->where('id', $idImagen);
+		$query = $this->db->get('ter_imagenes');
+
+		return $query->row_array();
+	}
+
+	public function deleteImagen($idImagen) {
+
+		$this->db->where('id', $idImagen);
+		$query = $this->db->delete('ter_imagenes');
+
+	}
+
+	public function deleteImagenTags($idImagen) {
+		
+		$this->db->where('id_imagen', $idImagen);
+		$query = $this->db->delete('ter_tags_on_imagenes');
+		
+	}
+
+
 
 /*================================================================
 	Categorías
@@ -161,6 +286,7 @@ class M_imagenes extends CI_Model {
 	} //#getTagByValue
 
 	public function getTagsToAlbum($idAlbum) {
+
 		$query = $this->db->query("SELECT  `ter_tags`.`value` FROM  `ter_tags` 
 									INNER JOIN  `ter_tags_on_albums` ON  `ter_tags_on_albums`.`id_tag` =  `ter_tags`.`id` 
 									WHERE  `ter_tags_on_albums`.`id_album` = ".$idAlbum);
@@ -170,6 +296,8 @@ class M_imagenes extends CI_Model {
 	} //#getTagsToAlbum
 
 	public function insertTag($value) {
+
+		$value = addslashes( strip_tags($value) );
 
 		$data = array('value' => $value );
 		$query = $this->db->insert('ter_tags', $data);
@@ -181,17 +309,60 @@ class M_imagenes extends CI_Model {
 
 	public function insertTagToAlbum($idAlbum, $idTag) {
 
-		$data = array('id_album' => $idAlbum, 'id_tag' => $idTag );
-		$query = $this->db->insert('ter_tags_on_albums', $data);
+		if ( !is_numeric($idAlbum) OR !is_numeric($idTag) ) {
+			die("Error en el envío del formulario");
+		} else {
+			$data = array('id_album' => $idAlbum, 'id_tag' => $idTag );
+			$query = $this->db->insert('ter_tags_on_albums', $data);
+		}
 
 	} //#insertTagToAlbum
 
 
 	public function deleteTagToAlbum($idAlbum) {
 
-		$this->db->where('id_album', $idAlbum);
-		$this->db->delete('ter_tags_on_albums');
+		if ( !is_numeric($idAlbum) ) {
+			die("Error en el envío del formulario");
+		} else {
+			$this->db->where('id_album', $idAlbum);
+			$this->db->delete('ter_tags_on_albums');
+		}
 
 	} //#insertTagToAlbum
+
+
+	public function getTagsToImagen($idImagen) {
+		
+		$query = $this->db->query("SELECT  `ter_tags`.`value` FROM  `ter_tags` 
+									INNER JOIN  `ter_tags_on_imagenes` ON  `ter_tags_on_imagenes`.`id_tag` =  `ter_tags`.`id` 
+									WHERE  `ter_tags_on_imagenes`.`id_imagen` = ".$idImagen);
+
+		return $query->result_array();
+
+	} //#getTagsToImagen
+
+
+	public function insertTagToImagen($idImagen, $idTag) {
+
+		if ( !is_numeric($idImagen) OR !is_numeric($idTag) ) {
+			die("Error en el envío del formulario");
+		} else {
+			$data = array('id_imagen' => $idImagen, 'id_tag' => $idTag );
+			$query = $this->db->insert('ter_tags_on_imagenes', $data);
+		}
+
+	} //#insertTagToImagen
+
+
+	public function deleteTagToImagen($idImagen) {
+
+		if ( !is_numeric($idImagen) ) {
+			die("Error en el envío del formulario");
+		} else {
+			$this->db->where('id_imagen', $idImagen);
+			$this->db->delete('ter_tags_on_imagenes');
+		}
+	}
+
 
 }
